Fix: Firefox did not re-enable Check Udpate button
Steps to reproduce: open CCU.IO interface in Firefox, check for updates and verify available version.
Click "reload" to open the page again, see the check button again, but still disabled. You would have to reenter the URL or close and reopen the tab/window to check for updates again.
This fixes this by unsetting the disabled attribute after hiding the button.